Add 'logic statement model' task/solver
This seems like a major change because the type of Op was changed
from taking a value/unary/binary function to taking a
unary/binary/ternary function, where the function's first argument is
an optional evaluation context. This is needed in order to evaluate a
logic expression, where a 0-arity operator may be a 'free variable',
meaning its value, when evaluated, depends on a specific interpretation.
Therefore all operators now have to specify an additional type 'c',
which states the type of context required to evaluate this operator.
Also all functions which will evaluate a tree further down, like
'findTree' or 'searchTree' take an instance of said context.

For Operators whose evaluation does *not* depend on any kind of context
the empty tuple `()` shall be taken as context type.